Fixed script so it could use preexisting build directories
The NXP setupenvironment script requires the conf directory to have
a sample conf file.  If the file is missing it does not find a
valid build directory and errors out.  Changed setup_beacon.sh to
copy that file to the build conf directory.  Also took a character
out of the bblayers.conf.sample file ( a '?') that should not be
there.

Signed-off-by: Richard Feliciano <RFeliciano@BeaconEmbedded.com>