Modified SDMA Kernal module loading
The SDMA kernel module was load order was causing BT
module errors due to the need to load it earlier. Modified
yocto build to add a file for the imx-sdma module.  This
means it load sooner and resolving BT errors see Jira Phoenix8-310

The recipe for the  ATF firmware code had to be updated

Signed-off-by: Richard Feliciano <RFeliciano@BeaconEmbedded.com>