Modified the custom Beacon meta layer to support Laird meta layer
The Laird meta layer is a vendor specific yocto layer to support
the Laird customized drivers and firmware. This layer is specific
to NXP because it supports Yocto Honister builds or later.

The meta-beacon-nxp layer was modified to build the Laird code
only if it is building imx8M Mini and Nano SOM BSPs

Also added code to do an automated download of Laird's latest
firmware.

Also updated recipes for symbolic links of firmware

Signed-off-by: Richard Feliciano <RFeliciano@BeaconEmbedded.com>