Don't enable Laird-CP Layer on imx8mp-beacon-kit
The Laird device is only used on imx8mm and imx8mn SOM's.
The imx8mp uses a different one, so including the Laird layer
isn't appropriate and might cause some conflicts.

Since the laird-cp layer may not be included, move the corresponding
bbappend files to a dynamic-layers directory so they don't throw
errors when the layer isn't included.