Missing LICENSE_CREATE_PACKAGE  in the sample local.conf file
The local.conf.sample is used to generate the local.conf.  It
is supposed to generate all license files by default.  This was
not happening because a parameter was left out and is now added.
LICENSE_CREATE_PACKAGE = "1"

See Jira [PHOENIX8-320] Adding License Text to Yocto Binary prior
to Releasing YOCTO Image

Signed-off-by: Richard Feliciano <RFeliciano@BeaconEmbedded.com>