layer: beacon: Disable fsl-rc-local
sysklogd used to set the loglevel to 1, so a script was added which
both reset the loglevel to 7 and also set the transparent huge page
to madvise.  Sysklog has already been disabled, and a patch for to
set madvise has already been submitted for the kernel defconfig file
which renders this script unnecessary.  Deleting fsl-rc-local
allows the proper loglevel to be set when passed from the command line
and also removes some splat when enabling and disabling services in
systemd.