systemd: networking: Set ethernet names for FEC and EQOS
The ethernet names for FEC and EQOS come up by default as eth0 and
eth1, but it's possible that with different kernel configurations
that these could enumerate in different orders. In order to guarantee
that the ethernet names correspond to their controller type, rename
the network connections to fec0 and eqos0 to correspond to their
controller type.

Fixes: GRIF-17