meta-beacon-nxp: beacon-imx8mp:  Add support for 2GB and Lite Variants
A new SOM has been created which has an 8MP Quad-Lite with 2GB of
LPDDR4. To support 2GB variants, and update to the bootloader is
necessary to properly configure the DRAM controller for this new
memory.  A new machine type was added to reference the updated
U-Boot configuration.

The Lite variant was hanging, so an update to the Kernel was neeeded.
To support the Lite variant which has the VPU stripped out, some
backports a device tree update were necessary to stop enabling
the VPU power domains when the VPU is not present.

Signed-off-by: Adam Ford <aford@beaconembedded.com>