imx-atf:  Update bbappend to point to right repo
The IMX-ATF pulled from NXP works fine for the Nano, Plus and 4G Mini,
but the 2G Mini appears to have a regression.  This requires a revert of
chunk of code from NXP. In order to do this, the bbappend needs to point
to the proper repo and the version needs to be correct.

With this ATF version, the older 2GB Mini configuration appears to work
correctly.

Signed-off-by: Adam Ford <aford@beaconembedded.com>